requirements.txt
pymysql
psycopg2
python-dotenv 